In the ASIOOutputReady() case, start hardware later
This is somewhat of a cosmetic change that shouldn't have practical
implications, but it does make the timeline clearer: the first two
bufferSwitch calls happen immediately at the beginning, then the
hardware is started, and and the third call only happens when one buffer
has been played back and recorded. This is more consistent with the
non-ASIOOutputReady case, and is also more consistent with ASIO SDK
docs, especially the timeine described in section 5. The previous
timeline would have an arbitrary amount of time to "start the hardware"
be inserted between the first two bufferSwitch calls, which was a bit
awkward.